Add Webpack for all modules
