+++ Task04/package.json
@@ -6,6 +6,8 @@
     "server:start": "node ./bin/www",
     "server:test": "cross-env NODE_ENV=tests jest",
     "client:build": "npx webpack",
+    "client:release": "cross-env NODE_ENV=production npx webpack",
+    "lint": "npx eslint .",
     "doc": "node docs/generate.js"
   },
   "dependencies": {
@@ -19,21 +21,24 @@
     "path-to-regexp": "^6.1.0",
     "pg": "^8.2.1",
     "pg-hstore": "^2.3.3",
-    "postgresql": "0.0.1",
     "sequelize": "^5.21.10",
     "swagger-ui-express": "^4.1.4",
     "umzug": "^2.3.0"
   },
   "devDependencies": {
+    "@egjs/hammerjs": "^2.0.17",
     "axios": "^0.19.2",
     "babel-loader": "^8.1.0",
+    "component-emitter": "^1.3.0",
     "cors": "^2.8.5",
     "cross-env": "^7.0.2",
     "eslint": "^7.0.0",
     "ify-loader": "^1.1.0",
     "jest": "^26.0.1",
     "jquery": "^3.5.1",
-    "plotly.js-dist": "^1.54.1",
+    "keycharm": "^0.3.1",
+    "moment": "^2.26.0",
+    "plotly.js": "^1.54.1",
     "rehype-document": "^5.0.0",
     "rehype-format": "^3.0.1",
     "rehype-highlight": "^4.0.0",
@@ -43,8 +48,13 @@
     "remark-rehype": "^7.0.0",
     "sequelize-cli": "^5.5.1",
     "sqlite3": "^4.2.0",
+    "timsort": "^0.3.0",
     "unified": "^9.0.0",
     "unist-util-visit": "^2.0.2",
+    "uuid": "^8.1.0",
+    "vis-data": "^6.5.3",
+    "vis-network": "^7.6.9",
+    "vis-util": "^4.0.2",
     "webpack": "^4.43.0",
     "webpack-cli": "^3.3.11"
   }
