+++ Task04/views/graphs/single.hbs
@@ -8,4 +8,4 @@
 
 </form>
 <div id="myChart" style="width:90%;height:70%;"></div>
-<script type="module" src="../js/filterSelectBoxAll.js"></script>
+<script src="../js/filterSelectBoxAll.js"></script>
