+++ Task04/views/graphs/sites.hbs
@@ -22,4 +22,4 @@
 	</select>
 </form>
 <div id="myChart" style="width:90%;height:70%;"></div>
-<script type="module" src="../js/filterSelectBoxInstGateCap.js"></script>
+<script src="../js/filterSelectBoxInstGateCap.js"></script>
